webSocket処理クラスが二つあったため削除
