--- Assets/MistNet/Runtime/Scripts/Core/WebSocketHandler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4bb1300f740e7354caf1c871505e8bcf
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
