+++ Assets/MistNet/Runtime/Scripts/Infra/WebSocketHandler.cs.meta
@@ -1,3 +1,11 @@
-﻿fileFormatVersion: 2
-guid: e4ae98cc2f6c45598c5a27f3839b1965
-timeCreated: 1744014464
+fileFormatVersion: 2
+guid: 4bb1300f740e7354caf1c871505e8bcf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
